From Peter Hrenka, "I implemented a free list reallocation scheme in VertexBufferObject::compileBuffer().
The offsets of newly added Arrays were not properly
calculated. This submission tries to find a
matching empty slot when the total size of
the VBO has not changed (e.g. when an array
is replaced by another array of the same size).


This fixes the overwriting issue that I showed in my posting
"Bug in VertexBufferObject::compileBuffer" on OSG-Users.
"


git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@9353 16af8721-9629-0410-8352-f15c8da7e697