From Mattias Helsing, "got some feedback from J-S on the packaging so I have changed the name of the top folder in each package to OpenSceneGraph-x.y.z
Noone else has had an opinion on whether compiler and it's version
should be part of the package filename on their platform/setup so I'll
leave it as is. That is - No action is taken for compilers other than
msvc but one can easily add this information to the cmake cache
regardless of platform.
"


git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@9528 16af8721-9629-0410-8352-f15c8da7e697