From David Callu, "I have a Segfault with xine plugin, when I load a video, then delete the XineStreamImage, then reload a video and finally delete the second XineStreamImage.
In src/osgPlugins/xine/video_out_rgb.c, many code is ASM code, and 'clear()' function is one of them.
If OSG is compiled without the flag COMPILE_ASSEMBLY (default behaviours) the clear() function is an empty
function and allocated memory is never initialized to 0. So a structure which contain pointer haven't its pointer set to NULL.
And when we need to delete this pointer, all go bad.

I join the fixed file."


git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@10806 16af8721-9629-0410-8352-f15c8da7e697