From Farshid Lashkari, "I've attached a small fix to osgUtil::SceneView so that is uses a scissor test when clearing the stencil buffer for stencil based stereo."
and

"I've added another small change for stencil based stereo, so please use this newer version. This newer version simplifies the calls to glOrtho and glRecti when drawing the stipple pattern. This change also happens to fix an issue where the stencil stereo would not work with certain viewport settings. I'm not exactly sure why this was happening, it might be a graphics driver issue, but either way I think the changes should be fine."



git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@10825 16af8721-9629-0410-8352-f15c8da7e697