Fixed Coverity reported issue.  Issue is benign as all the member variables are initialized in after construction, but I've moved this initialization into the constructor to make the code more managable.
CID 11686: Uninitialized scalar field (UNINIT_CTOR)
Non-static class member cnt is not initialized in this constructor nor in any functions that it calls.
Non-static class member maxcnt is not initialized in this constructor nor in any functions that it calls.
Non-static class member phase is not initialized in this constructor nor in any functions that it calls.



git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@12326 16af8721-9629-0410-8352-f15c8da7e697