Fixed Coverity reported issue.
CID 11403: Logically dead code (DEADCODE)
After this line, the value of "needspace" is equal to 0.
Assigning: "needspace" = "false".

Technically the report is correct, but the code is just debugging code that was there just in case
the code needed to be reviewed. This code is long since through it's debugging stage so the code
is no longer really worth keeping so I've removed it for cleanness sake.



git-svn-id: http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk@12329 16af8721-9629-0410-8352-f15c8da7e697